Module_6/API/covid19-regression/src/model-api/requirements.txt
altair==4.0.1
appnope==0.1.0
astroid==2.3.3
astropy==4.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cmdstanpy==0.4.0
confuse==1.0.0
convertdate==2.2.0
cycler==0.10.0
Cython==0.29.15
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
ephem==3.7.7.1
et-xmlfile==1.0.1
fbprophet==0.6
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Jsonpify==1.5.0
holidays==0.10.1
htmlmin==0.1.12
idna==2.9
importlib-metadata==1.4.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.2
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==2.0.1
jupyterlab-server==1.0.7
kaggle==1.5.6
kiwisolver==1.1.0
lab==5.5
lazy-object-proxy==1.4.3
llvmlite==0.31.0
LunarCalendar==0.0.9
MarkupSafe==1.1.1
matplotlib==3.2.0
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
more-itertools==8.1.0
nbconvert==5.6.1
nbformat==5.0.3
networkx==2.4
notebook==6.0.2
numba==0.47.0
numpy==1.18.1
openpyxl==3.0.3
packaging==20.0
pandas==0.25.3
pandas-profiling==2.5.0
pandocfilters==1.4.2
parso==0.5.2
patsy==0.5.1
pexpect==4.7.0
phik==0.9.10
pickleshare==0.7.5
plotly==4.5.4
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.2
ptyprocess==0.6.0
py==1.8.1
Pygments==2.5.2
pylint==2.4.4
PyMeeus==0.3.7
pyparsing==2.4.6
PyQt5==5.14.1
PyQt5-sip==12.7.1
pyrsistent==0.15.7
pystan==2.19.1.1
pytest==5.3.2
pytest-pylint==0.14.1
python-dateutil==2.8.1
python-decouple==3.3
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.7.1
QtPy==1.9.0
requests==2.23.0
retrying==1.3.3
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
setuptools-git==1.2
simplejson==3.17.0
six==1.13.0
statsmodels==0.11.1
tangled-up-in-unicode==0.0.4
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
toolz==0.10.0
tornado==6.0.3
tqdm==4.43.0
traitlets==4.3.3
txt2tags==3.7
typed-ast==1.4.1
urllib3==1.25.8
vega-datasets==0.8.0
visions==0.2.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==1.0.0
